@@ -5,42 +5,55 @@
   "scripts": {
     "serve": "vue-cli-service serve",
     "build": "vue-cli-service build && rm -rf docs && mv dist docs && cp CNAME docs/CNAME",
-    "distribute": "vue-cli-service serve --mode development",
-    "update": "./scripts/update.sh"
+    "distribute": "vue-cli-service serve --mode development"
   },
   "dependencies": {
-    "@types/node": "^13.13.5",
+    "@tailwindcss/forms": "^0.2.1",
+    "@types/node": "^14.14.31",
     "axios": "^0.21.1",
-    "moment": "^2.25.3",
-    "pug": "^2.0.3",
-    "pug-plain-loader": "^1.0.0",
-    "randomcolor": "^0.5.4",
-    "vue": "^2.6.11",
-    "vue-class-component": "^7.2.3",
+    "moment": "^2.29.1",
+    "pug": "^3.0.0",
+    "pug-plain-loader": "^1.1.0",
+    "randomcolor": "^0.6.2",
+    "tailwindcss-filters": "2.x",
+    "vue": "^2.6.12",
+    "vue-class-component": "^7.2.6",
+    "vue-click-outside": "^1.1.0",
     "vue-cookie-law": "^1.13.3",
-    "vue-i18n": "^8.17.5",
-    "vue-property-decorator": "^8.4.2",
-    "vue-router": "^3.1.6",
-    "vuetify": "^2.2.28",
-    "vuex": "^3.4.0",
-    "vuex-persistedstate": "^3.0.1"
+    "vue-i18n": "^8.22.4",
+    "vue-property-decorator": "^9.1.2",
+    "vue-router": "^3.5.1",
+    "vuetify": "^2.4.4",
+    "vuex": "^3.6.2",
+    "vuex-class": "^0.3.2",
+    "vuex-module-decorators": "^1.0.1",
+    "vuex-persistedstate": "^4.0.0-beta.3"
   },
   "devDependencies": {
-    "@vue/cli-plugin-typescript": "^4.3.1",
-    "@vue/cli-service": "^4.3.1",
-    "sass": "^1.26.5",
-    "sass-loader": "^8.0.2",
-    "typescript": "^3.9.2",
-    "vue-template-compiler": "^2.6.11"
-  },
-  "postcss": {
-    "plugins": {
-      "autoprefixer": {}
-    }
+    "@prettier/plugin-pug": "^1.13.4",
+    "@tailwindcss/postcss7-compat": "^2.0.3",
+    "@vue/cli-plugin-typescript": "^4.5.11",
+    "@vue/cli-service": "^4.5.11",
+    "autoprefixer": "^9",
+    "postcss": "^7",
+    "prettier": "prettier/prettier#cd2b8a32b5ef77eb0047e7517b1ea22d605a094a",
+    "sass": "^1.32.8",
+    "sass-loader": "^10.1.1",
+    "tailwindcss": "npm:@tailwindcss/postcss7-compat",
+    "typescript": "^4.1.5",
+    "vue-template-compiler": "^2.6.12"
   },
   "browserslist": [
     "> 1%",
     "last 2 versions",
     "not ie <= 8"
-  ]
+  ],
+  "prettier": {
+    "pluginSearchDirs": [
+      "."
+    ],
+    "semi": false,
+    "singleQuote": true,
+    "trailingComma": "es5"
+  }
 }
