@@ -1,270 +1,387 @@
 <template lang="pug">
-  v-card.mb-2
-    v-card-title
-      v-btn.mr-1(
-        v-if='admin'
-        color='red'
-        small
-        icon
-        @click='deleteLocalization(localization.key)'
+.card
+  .card__head
+    .card__icons
+      Icon(
+        v-if='isAdmin',
+        @click='deleteLocalization(localization.key)',
+        :loading='loading'
+      ) 
+        img(src='../assets/icons/close.svg', alt='Delete') 
+      Icon(
+        v-if='isAdmin',
+        @click='selectOrDeleteVariantsEnabled = !selectOrDeleteVariantsEnabled',
         :loading='loading'
       )
-        v-icon(small) delete
-      v-btn.mr-1(
-        v-if='admin'
-        small
-        icon
-        @click='select = !select'
-        :loading='loading'
-        :class='select ? "green darken-2" : ""'
-      )
-        v-icon(small) edit
-      v-btn.mr-1(
-        v-if='!edit'
-        small
-        icon
-        @click='edit = true'
-      )
-        v-icon(small) add
-      v-btn.mr-1(
-        v-else
-        small
-        icon
-        @click='edit = false'
-        color='red'
-      )
-        v-icon(small) clear
-      span.mx-1 {{localization.key}}
-      v-chip.mx-1(
-        dark
-        small
-        v-for='tag in localization.tags'
-        :key='tag'
-        :color='$store.state.colors[tag]'
-      )
-        span {{tag}}
-        .ml-2(
-          small
-          v-if='admin'
-          @click='deleteLocalizationTag(localization.key, tag)'
-          :disabled='loading'
-        )
-          v-icon(small v-if='!loading') close
-          span(v-else) 🤔
-      v-chip.mx-1.px-1(
-        v-if='!$store.state.viewedItems[localization._id]'
-        dark
-        small
-        @mouseover='setViewedItem(localization._id)'
-        color='primary'
-      ) {{$t('new')}}
-      v-chip.mx-1.px-1(
-        dark
-        small
-        v-if='admin'
-        :color='addTag ? "green darken-2" : ""'
-        @click='addTag = !addTag'
-      )
-        v-icon(small v-if='!loading') {{addTag ? 'close' : 'add'}}
-      v-text-field.mx-1.mb-0.mt-4.px-0.py-0(
-        :label='$t("tag.new")'
-        v-if='addTag'
-        clearable
-        rows='1'
-        auto-grow
-        no-resize
-        compact
-        v-model='addTagText'
-        :append-outer-icon="!!addTagText ? 'send' : undefined"
-        @click:append-outer='saveNewTag'
-        :disabled='loading'
-      )
-    v-card-text
-      div(v-if='edit')
-        v-textarea.mb-1.mt-0.pt-0(
-          :label='$t("add.text")'
-          clearable
-          rows='1'
-          :rules='textRules'
-          auto-grow
-          no-resize
-          compact
-          v-model='text'
-        )
-        .d-flex
-          v-select.mb-1.mt-0.pt-0(
-            :label='$t("add.language")'
-            :items='languages'
-            :rules='languageRules'
-            v-model='language'
-          )
-          v-btn.ml-2(
-            color='primary'
-            :disabled='!text || !language'
-            @click='save(localization.key)'
+        img(src='../assets/icons/edit.svg', alt='Edit') 
+      Icon(@click='toggleAddVariantEnabled', :loading='loading') 
+        img(src='../assets/icons/add.svg', alt='Add') 
+    h2.card__title {{ localization.key }}
+    .card__chips
+      .flex.items-center.space-x-2.flex-wrap
+        Chip(
+          v-for='tag in localization.tags',
+          :key='tag',
+          :color='colors[tag]',
+          small,
+          selected,
+          inactive
+        ) 
+          span {{ tag }}
+          Icon(
+            v-if='isAdmin',
+            @click='deleteTag(localization.key, tag)',
             :loading='loading'
-          ) {{$t("add.save")}}
-      div(v-if='select')
-        v-chip.px-1(
-          dark
-          x-small
-          color='red'
-          @click='deleteVariants(localization.key)'
-          :loading='loading'
-        )
-          v-icon(x-small color='white') delete
-        v-chip.px-1(
-          dark
-          x-small
-          color='green'
-          @click='selectVariants(localization.key)'
-          :loading='loading'
-        )
-          v-icon(x-small color='white') done
-      div(v-for='variant in localization.variants')
-        .d-flex.direction-row
-          v-checkbox(
-            v-if='select'
-            v-model='selected[variant._id]'
           )
-          Variant(
-            :variant='variant'
-            :loadData='loadData'
-            :admin='admin'
-            :localization='localization'
-            :select='select'
-          )
+            img(src='../assets/icons/x.svg', width=15) 
+        Chip(
+          isNew,
+          small,
+          v-if='!viewedItems[localization._id]',
+          @click='setViewedProxy'
+        ) {{ $t("new") }}
+        Icon(v-if='isAdmin && !loading', @click='toggleAddTagEnabled') 
+          img(src='../assets/icons/add.svg', alt='Add')
+  .card__controls
+    .input-group(v-if='addTagEnabled && isAdmin')
+      Input(type='text', :label='$t("tag.new")', v-model='addTagText')
+      Button(:inactive='!addTagText', @click='addTag') {{ $t("add.save") }}
+    .input-group(v-if='addVariantEnabled')
+      Input(type='text', :label='$t("add.text")', v-model='addVariantText')
+      Select(
+        @click='setVariantLanguage',
+        :items='languages',
+        :label='$t("add.language")'
+      )
+      Button(
+        :inactive='!(addVariantText && addVariantLanguage)',
+        @click='addVariant'
+      ) {{ $t("add.save") }}
+  div(v-if='selectOrDeleteVariantsEnabled')
+    .flex.items-center.pl-5.pb-5
+      Chip(@click='deleteVariants', :loading='loading') Delete
+      Chip(@click='selectVariants', :loading='loading') Mark as done
+  .card__body(
+    v-for='variant in localization.variants',
+    :class='selectedVariants[variant._id] ? "card__body--selected" : ""'
+  )
+    .flex.flex-row.space-x-3.items-center
+      div(v-if='selectOrDeleteVariantsEnabled')
+        Button(
+          huge,
+          v-model='selectedVariants[variant._id]',
+          @click='addToSelected(variant._id)'
+        ) {{ selectedVariants[variant._id] ? "Unselect" : "Select" }}
+      VariantView(
+        :variant='variant',
+        :localization='localization',
+        :selected='selectedVariants[variant._id]',
+        :selectOrDeleteVariantsEnabled='selectOrDeleteVariantsEnabled'
+      )
 </template>
 
 <script lang="ts">
 import Vue from 'vue'
 import Component from 'vue-class-component'
-import * as store from '../plugins/store'
-import { i18n } from '../plugins/i18n'
-import * as api from '../utils/api'
-import Variant from './Variant.vue'
+import { i18n } from '@/plugins/i18n'
+import * as api from '@/utils/api'
+import VariantView from '@/components/Variant.vue'
+import { namespace } from 'vuex-class'
+import { ColorsMap } from '@/models/ColorsMap'
+import { Variant } from '@/models/Variant'
+import { ViewedItems } from '@/models/ViewedItems'
+
+const SnackbarStore = namespace('SnackbarStore')
+const AppStore = namespace('AppStore')
+const DataStore = namespace('DataStore')
 
 @Component({
   props: {
     localization: Object,
-    languages: Array,
-    loadData: Function,
-    admin: Boolean,
   },
   components: {
-    Variant,
+    VariantView,
   },
 })
 export default class LocalizationCard extends Vue {
-  text = ''
-  language = ''
+  @AppStore.State username!: string
+  @AppStore.State isAdmin!: boolean
+  @DataStore.State colors!: ColorsMap
+  @DataStore.State languages!: string[]
+  @DataStore.State viewedItems!: ViewedItems
+
+  @SnackbarStore.Mutation setSnackbarError!: (error: string) => void
+  @DataStore.Mutation setViewedItem!: (id: string) => void
+  @DataStore.Mutation removeLocalization!: (key: string) => void
+  @DataStore.Mutation removeLocalizationTag!: (options: {
+    key: string
+    tag: string
+  }) => void
+  @DataStore.Mutation addLocalizationTag!: (options: {
+    key: string
+    tag: string
+  }) => void
+  @DataStore.Mutation addLocalizationVariant!: (options: {
+    key: string
+    variant: Variant
+  }) => void
+  @DataStore.Mutation deleteLocalizationVariants!: (options: {
+    key: string
+    variantIds: string[]
+  }) => void
+  @DataStore.Mutation selectLocalizationVariants!: (options: {
+    key: string
+    variantIds: string[]
+  }) => void
+  @DataStore.Mutation refreshLocalizations!: () => void
+
+  addVariantText = ''
+  addVariantLanguage = ''
   addTagText = ''
 
-  edit = false
-  select = false
-  addTag = false
+  addVariantEnabled = false
+  selectOrDeleteVariantsEnabled = false
+  addTagEnabled = false
 
   textRules = [(v: any) => !!(v || '').trim() || i18n.t('errors.textLength')]
   languageRules = [(v: any) => !!(v || '').trim() || i18n.t('errors.language')]
 
   loading = false
 
-  selected = {} as any
-
-  async save(key: string) {
-    this.loading = true
-    try {
-      await api.postVariant(key, this.text, this.language, store.username())
-      this.$props.loadData()
-      this.text = ''
-      this.edit = false
-    } catch (err) {
-      store.setSnackbarError(err.response.data)
-    } finally {
-      this.loading = false
-    }
-  }
-
-  async saveNewTag() {
-    this.loading = true
-    try {
-      if (this.$props.localization.tags.indexOf(this.addTagText) < 0) {
-        await api.addLocalizationTag(
-          this.$props.localization.key,
-          this.addTagText
-        )
-        this.$props.localization.tags.push(this.addTagText)
-      }
-      this.addTagText = ''
-    } catch (err) {
-      store.setSnackbarError(err.response.data)
-    } finally {
-      this.loading = false
-    }
-  }
+  selectedVariants = {} as any
 
   async deleteLocalization(key: string) {
-    this.loading = true
-    try {
+    this.performRequest(async () => {
       await api.deleteLocalization(key)
-      this.$props.loadData()
-    } catch (err) {
-      store.setSnackbarError(err.response.data)
-    } finally {
-      this.loading = false
+      this.removeLocalization(key)
+    })
+  }
+
+  setVariantLanguage(variant: string) {
+    this.addVariantLanguage = variant;
+  }
+
+  addToSelected(id: string) {
+    if (this.selectedVariants[id]) {
+      this.selectedVariants[id] = undefined
+    } else {
+      this.selectedVariants = { ...this.selectedVariants, [id]: true }
     }
   }
 
-  async deleteLocalizationTag(key: string, tag: string) {
-    this.loading = true
-    try {
-      await api.deleteLocalizationTag(key, tag)
-      this.$props.localization.tags = this.$props.localization.tags.filter(
-        (t: string) => t !== tag
-      )
-    } catch (err) {
-      store.setSnackbarError(err.response.data)
-    } finally {
-      this.loading = false
-    }
-  }
-
-  async deleteVariants(key: string) {
-    this.loading = true
-    try {
-      await api.deleteVariants(
+  async addVariant() {
+    const key = this.$props.localization.key
+    this.performRequest(async () => {
+      const variant = await api.postVariant(
         key,
-        Object.keys(this.selected).filter((k) => !!this.selected[k])
+        this.addVariantText,
+        this.addVariantLanguage,
+        this.username
       )
-      this.selected = {}
-      this.select = false
-      this.$props.loadData()
-    } catch (err) {
-      store.setSnackbarError(err.response.data)
-    } finally {
-      this.loading = false
-    }
+      this.addLocalizationVariant({ key, variant })
+      this.resetAddVariant()
+    })
   }
 
-  async selectVariants(key: string) {
+  async addTag() {
+    const tag = this.addTagText
+    const key = this.$props.localization.key
+    this.performRequest(async () => {
+      if (!this.$props.localization.tags.includes(tag)) {
+        await api.addLocalizationTag(key, tag)
+        this.addLocalizationTag({ key, tag })
+        this.refreshLocalizations()
+      }
+      this.resetAddTag()
+    })
+  }
+
+  async deleteTag(key: string, tag: string) {
+    this.performRequest(async () => {
+      if (this.$props.localization.tags.includes(tag)) {
+        await api.deleteLocalizationTag(key, tag)
+        this.removeLocalizationTag({ key, tag })
+        this.refreshLocalizations()
+      }
+    })
+  }
+
+  async deleteVariants() {
+    const ids = Object.keys(this.selectedVariants).filter(
+      (k) => !!this.selectedVariants[k]
+    )
+    const key = this.$props.localization.key
+    this.performRequest(async () => {
+      await api.deleteVariants(key, ids)
+      this.deleteLocalizationVariants({ key, variantIds: ids })
+      this.selectedVariants = {}
+      this.selectOrDeleteVariantsEnabled = false
+    })
+  }
+
+  selectVariants() {
+    const ids = Object.keys(this.selectedVariants).filter(
+      (k) => !!this.selectedVariants[k]
+    )
+    const key = this.$props.localization.key
+    this.performRequest(async () => {
+      await api.selectVariants(key, ids)
+      this.selectLocalizationVariants({ key, variantIds: ids })
+      this.selectedVariants = {}
+      this.selectOrDeleteVariantsEnabled = false
+    })
+  }
+
+  async performRequest(requestFunction: () => Promise<unknown>) {
     this.loading = true
     try {
-      await api.selectVariants(
-        key,
-        Object.keys(this.selected).filter((k) => !!this.selected[k])
-      )
-      this.selected = {}
-      this.edit = false
-      this.$props.loadData()
+      await requestFunction()
     } catch (err) {
-      store.setSnackbarError(err.response.data)
+      console.error(err)
+      this.setSnackbarError(err.response?.data || JSON.stringify(err))
     } finally {
       this.loading = false
     }
   }
 
-  setViewedItem(id: string) {
-    store.setViewedItem(id)
+  setViewedProxy() {
+    this.setViewedItem(this.$props.localization._id)
+    this.refreshLocalizations()
+  }
+
+  toggleAddTagEnabled() {
+    this.addTagEnabled = !this.addTagEnabled
+    if (!this.addTagEnabled) {
+      this.resetAddTag()
+    }
+  }
+
+  resetAddTag() {
+    this.addTagText = ''
+    this.addTagEnabled = false
+  }
+
+  toggleAddVariantEnabled() {
+    this.addVariantEnabled = !this.addVariantEnabled
+    if (!this.addVariantEnabled) {
+      this.resetAddVariant()
+    }
+  }
+
+  resetAddVariant() {
+    this.addVariantText = ''
+    this.addVariantLanguage = ''
+    this.addVariantEnabled = false
   }
 }
 </script>
+
+<style lang="scss">
+.card {
+  @apply transition;
+  @apply mb-5;
+  @apply border;
+  @apply border-back-silver;
+  @apply rounded-lg;
+  @apply shadow-sm;
+  @apply bg-white;
+
+  &__head {
+    @apply p-5;
+    @apply flex;
+    @apply justify-start;
+    @apply space-x-0;
+    @apply md_space-x-5;
+    @apply flex-col;
+    @apply md_flex-row;
+    @apply items-center;
+    @apply space-y-2;
+    @apply md_space-y-0;
+  }
+
+  &__title {
+    @apply font-medium;
+    @apply text-text-dark;
+    @apply text-lg;
+    @apply break-all;
+    @apply place-self-start;
+  }
+
+  &__variant {
+    @apply w-full;
+  }
+
+  &__chips {
+    @apply place-self-start;
+  }
+
+  &__controls {
+    @apply flex;
+    @apply flex-col;
+  }
+
+  &__icons {
+    @apply flex;
+    @apply items-center;
+    @apply place-self-start;
+    @apply space-x-1;
+  }
+
+  &__body {
+    @apply transition;
+    @apply border-t-2;
+    @apply border-back-silver;
+    @apply p-5;
+
+    &--selected {
+      @apply shadow-inner;
+      @apply opacity-70;
+    }
+  }
+
+  &__comments {
+    @apply mt-5;
+  }
+}
+
+.select {
+  min-width: 150px;
+  @apply bg-transparent;
+  @apply text-xl;
+  @apply cursor-pointer;
+  @apply text-text-silver;
+  @apply font-sans;
+  @apply font-medium;
+  @apply transition;
+  @apply block;
+  @apply p-3;
+  @apply md_p-4;
+  @apply rounded-2xl;
+  @apply border-2;
+  @apply border-back-silver;
+  @apply focus_border-primary-blue;
+  @apply focus_ring-0;
+  @apply placeholder-text-light;
+}
+
+.dark .select {
+  @apply border-dark-card-border;
+  @apply focus_border-primary-blue;
+  @apply text-text-milk;
+  @apply placeholder-text-dark;
+}
+
+.dark .card {
+  @apply bg-black;
+  @apply border-dark-card-border;
+  @apply shadow-none;
+
+  &__title {
+    @apply text-text-milk;
+  }
+
+  &__body {
+    @apply border-dark-card-border;
+  }
+}
+</style>
