@@ -1,27 +1,33 @@
 <template lang="pug">
-  v-text-field(
-    :append-icon="show ? 'mdi-eye' : 'mdi-eye-off'"
-    :label="$t('admin.password')"
-    v-model='password'
-    :type='show ? "text" : "password"'
-    @click:append="show = !show"
-  )
+Input(
+  type='password',
+  :label='$t("admin.password")',
+  v-model='passwordProxy',
+  icon='password'
+)
 </template>
 
 <script lang="ts">
 import Vue from 'vue'
 import Component from 'vue-class-component'
-import * as store from '../plugins/store'
+import { namespace } from 'vuex-class'
+
+const AppStore = namespace('AppStore')
 
 @Component
 export default class PasswordInput extends Vue {
+  @AppStore.State password!: string
+  @AppStore.Mutation setPassword!: (password: string) => void
+
   show = false
 
-  get password() {
-    return store.password()
+  get passwordProxy() {
+    return this.password
   }
-  set password(newPassword: string | undefined) {
-    store.setPassword(newPassword)
+
+  set passwordProxy(newPassword: string) {
+    this.setPassword(newPassword)
   }
+
 }
 </script>
