@@ -1,9 +1,8 @@
 import Vue from 'vue'
 import Router from 'vue-router'
-import Home from '../views/Home.vue'
-import Admin from '../views/Admin.vue'
-import Code from '../views/Code.vue'
-import NotFound from '../views/NotFound.vue'
+import Home from '@/views/Home.vue'
+import Code from '@/views/Code.vue'
+import NotFound from '@/views/NotFound.vue'
 
 Vue.use(Router)
 
@@ -14,11 +13,6 @@ const router = new Router({
       path: '/',
       name: 'home',
       component: Home,
-    },
-    {
-      path: '/admin',
-      name: 'admin',
-      component: Admin,
     },
     {
       path: '/code',
