@@ -1,10 +1,37 @@
+import { Comment } from '@/models/Comment'
+import { Variant } from '@/models/Variant'
+import { Localization } from '@/models/Localization'
 import axios from 'axios'
-import * as store from '../plugins/store'
+import store from '@/store'
 
-const base = process.env.VUE_APP_API
+const base = process.env.VUE_APP_API || '/api/'
+
+function convertCreatedAt(item: any) {
+  item.createdAt = new Date(item.createdAt)
+  return item
+}
 
 export async function getLocalizations() {
   return (await axios.get(`${base}/localizations`)).data
+    .map(convertCreatedAt)
+    .map((localization: any) => {
+      localization.variants.forEach((variant: any) => {
+        variant.comments = variant.comments.map(convertCreatedAt)
+      })
+      return localization
+    })
+    .map((localization: Localization) => {
+      localization.variants = localization.variants.sort((a: any, b: any) => {
+        if (a.selected && !b.selected) {
+          return -1
+        } else if (!a.selected && b.selected) {
+          return 1
+        } else {
+          return 0
+        }
+      })
+      return localization
+    }) as Localization[]
 }
 
 export async function postVariant(
@@ -20,7 +47,7 @@ export async function postVariant(
       language,
       username,
     })
-  ).data
+  ).data as Variant
 }
 
 export async function selectVariant(key: string, id: string) {
@@ -28,7 +55,7 @@ export async function selectVariant(key: string, id: string) {
     await axios.post(`${base}/localizations/select`, {
       key,
       id,
-      password: store.password(),
+      password: (store as any).state.AppStore.password,
     })
   ).data
 }
@@ -38,7 +65,7 @@ export async function deleteVariant(key: string, id: string) {
     await axios.post(`${base}/localizations/delete`, {
       key,
       id,
-      password: store.password(),
+      password: (store as any).state.AppStore.password,
     })
   ).data
 }
@@ -48,7 +75,7 @@ export async function selectVariants(key: string, ids: string[]) {
     await axios.post(`${base}/localizations/select`, {
       key,
       ids,
-      password: store.password(),
+      password: (store as any).state.AppStore.password,
     })
   ).data
 }
@@ -58,7 +85,7 @@ export async function deleteVariants(key: string, ids: string[]) {
     await axios.post(`${base}/localizations/delete`, {
       key,
       ids,
-      password: store.password(),
+      password: (store as any).state.AppStore.password,
     })
   ).data
 }
@@ -69,7 +96,7 @@ export async function editVariant(key: string, id: string, text: string) {
       key,
       id,
       text,
-      password: store.password(),
+      password: (store as any).state.AppStore.password,
     })
   ).data
 }
@@ -114,7 +141,7 @@ export async function deleteLocalization(key: string) {
   return (
     await axios.post(`${base}/localizations/localization/delete`, {
       key,
-      password: store.password(),
+      password: (store as any).state.AppStore.password,
     })
   ).data
 }
@@ -124,7 +151,7 @@ export async function deleteLocalizationTag(key: string, tag: string) {
     await axios.post(`${base}/localizations/localization/deleteTag`, {
       key,
       tag,
-      password: store.password(),
+      password: (store as any).state.AppStore.password,
     })
   ).data
 }
@@ -134,7 +161,7 @@ export async function addLocalizationTag(key: string, tag: string) {
     await axios.post(`${base}/localizations/localization/addTag`, {
       key,
       tag,
-      password: store.password(),
+      password: (store as any).state.AppStore.password,
     })
   ).data
 }
@@ -146,12 +173,12 @@ export async function leaveCommentToVariant(
 ) {
   return (
     await axios.post(`${base}/localizations/comment`, {
-      username: store.username(),
+      username: (store as any).state.AppStore.username,
       text,
       key,
       id,
     })
-  ).data
+  ).data as Comment
 }
 
 export async function deleteCommentToVariant(
@@ -164,7 +191,7 @@ export async function deleteCommentToVariant(
       key,
       id,
       commentId,
-      password: store.password(),
+      password: (store as any).state.AppStore.password,
     })
   ).data
 }
