@@ -1,5 +1,6 @@
 module.exports = {
   devServer: {
     disableHostCheck: true
-  }
+  },
+  productionSourceMap: false
 }
